Use enum init instead of static func for the streamType
