+++ Tests/PlayerTests/PlayerTests.swift
@@ -26,35 +26,4 @@ final class PlayerTests: XCTestCase {
         }
         expect(weakPlayer).to(beNil())
     }
-
-    func testStreamTypeWhenTimeRangeAndItemDurationAreInvalid() {
-        let streamType = Player.streamType(for: .invalid, itemDuration: .invalid)
-        expect(streamType).to(equal(.unknown))
-    }
-
-    func testStreamTypeWhenTimeRangeIsEmpty() {
-        let streamType = Player.streamType(for: .zero, itemDuration: .indefinite)
-        expect(streamType).to(equal(.live))
-    }
-
-    func testStreamTypeWhenItemDurationIsIndefinite() {
-        let start = CMTime(value: 0, timescale: 1)
-        let end = CMTime(value: 1, timescale: 1)
-        let streamType = Player.streamType(for: .init(start: start, end: end), itemDuration: .indefinite)
-        expect(streamType).to(equal(.dvr))
-    }
-
-    func testStreamTypeWhenTimeRangeIsNotEmptyAndItemDurationIsZero() {
-        let start = CMTime(value: 0, timescale: 1)
-        let end = CMTime(value: 1, timescale: 1)
-        let streamType = Player.streamType(for: .init(start: start, end: end), itemDuration: .zero)
-        expect(streamType).to(equal(.live))
-    }
-
-    func testStreamTypeWhenTimeRangeIsNotEmptyAndItemDurationIsDefinedAndNotZero() {
-        let start = CMTime(value: 0, timescale: 1)
-        let end = CMTime(value: 1, timescale: 1)
-        let streamType = Player.streamType(for: .init(start: start, end: end), itemDuration: end)
-        expect(streamType).to(equal(.onDemand))
-    }
 }
